Platform generation

+++ Starving Steve/src/starter/Block.java
@@ -13,7 +13,7 @@ public class Block {
 		this.y = y;
 	}
 	
-	public void setLocation(int x, int y)
+	public void setLocation(double x, double y)
 	{
 		this.x = x;
 		this.y = y;
